chore: break up bundle tests
