+++ .github/workflows/parallel-tests.yaml
@@ -32,6 +32,9 @@ jobs:
   test-bundle:
     runs-on: ubuntu-latest
     needs: build
+    strategy:
+      matrix:
+        test: [bundle, deploy-flags]
     steps:
       - name: Checkout
         uses: actions/checkout@692973e3d937129bcbf40652eb9f2f61becf3332 # v4.1.7
@@ -43,14 +46,13 @@ jobs:
 
       - name: Run e2e bundle tests
         run: |
-          build/uds run test:bundle
+          build/uds run test:${{ matrix.test }}
 
       - name: Save logs
         if: always()
         uses: defenseunicorns/uds-common/.github/actions/save-logs@172a905901cb9bd76d096d2850bf31af5c5a4fa1 # v0.8.0
         with:
           suffix: build
-
 
   test-dev:
     runs-on: ubuntu-latest
@@ -74,7 +76,6 @@ jobs:
         with:
           suffix: test-dev
 
-
   test-variables:
     runs-on: ubuntu-latest
     needs: build
@@ -97,7 +98,6 @@ jobs:
         with:
           suffix: test-variables
 
-
   test-optional-bundle:
     runs-on: ubuntu-latest
     needs: build
@@ -119,7 +119,6 @@ jobs:
         uses: defenseunicorns/uds-common/.github/actions/save-logs@172a905901cb9bd76d096d2850bf31af5c5a4fa1 # v0.8.0
         with:
           suffix: test-optional-bundle
-
 
   test-vendor:
     runs-on: ubuntu-latest
@@ -151,7 +150,6 @@ jobs:
         uses: defenseunicorns/uds-common/.github/actions/save-logs@172a905901cb9bd76d096d2850bf31af5c5a4fa1 # v0.8.0
         with:
           suffix: test-vendor
-
 
   test-engine:
     runs-on: ubuntu-latest
